+++ ribs/emitters/_gradient_arborescence_emitter.py
@@ -312,7 +312,7 @@ class GradientArborescenceEmitter(EmitterBase):
         raise ValueError(f"Invalid restart_rule {self._restart_rule}")
 
     def tell_dqd(self, solution, objective, measures, jacobian, status_batch,
-                 value_batch):
+                 value_batch, **fields):
         """Gives the emitter results from evaluating the gradient of the
         solutions.
 
@@ -334,6 +334,9 @@ class GradientArborescenceEmitter(EmitterBase):
             value_batch (array-like): 1d array of floats returned by a series
                 of calls to archive's :meth:`add()` method. for what these
                 floats represent, refer to :meth:`ribs.archives.add()`.
+            fields (keyword arguments): Additional data for each solution. Each
+                argument should be an array with batch_size as the first
+                dimension.
         """
         data, add_info, jacobian = validate_batch(  # pylint: disable = unused-variable
             self.archive,
@@ -341,6 +344,7 @@ class GradientArborescenceEmitter(EmitterBase):
                 "solution": solution,
                 "objective": objective,
                 "measures": measures,
+                **fields,
             },
             {
                 "status": status_batch,
@@ -355,7 +359,8 @@ class GradientArborescenceEmitter(EmitterBase):
             jacobian /= norms
         self._jacobian_batch = jacobian
 
-    def tell(self, solution, objective, measures, status_batch, value_batch):
+    def tell(self, solution, objective, measures, status_batch, value_batch,
+             **fields):
         """Gives the emitter results from evaluating solutions.
 
         The solutions are ranked based on the `rank()` function defined by
@@ -374,6 +379,9 @@ class GradientArborescenceEmitter(EmitterBase):
             value_batch (array-like): 1d array of floats returned by a series
                 of calls to archive's :meth:`add()` method. for what these
                 floats represent, refer to :meth:`ribs.archives.add()`.
+            fields (keyword arguments): Additional data for each solution. Each
+                argument should be an array with batch_size as the first
+                dimension.
         Raises:
             RuntimeError: This method was called without first passing gradients
                 with calls to ask_dqd() and tell_dqd().
@@ -384,6 +392,7 @@ class GradientArborescenceEmitter(EmitterBase):
                 "solution": solution,
                 "objective": objective,
                 "measures": measures,
+                **fields,
             },
             {
                 "status": status_batch,